+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Procedure when bumping the version number:
 
 ## Unreleased
 
+### Fixed
+- Width measurements of ASCII control characters
+
 ## v0.2.2 - 2024-01-14
 
 ### Fixed